Stylo: Update bindings for ServoType

This is needed for https://bugzilla.mozilla.org/show_bug.cgi?id=1348873

---
<!-- Thank you for contributing to Servo! Please replace each `[ ]` by `[X]` when the step is complete, and replace `__` with appropriate data: -->
- [X] `./mach build -d` does not report any errors
- [X] `./mach test-tidy` does not report any errors
- [X] These changes fix [Bug 1348873](https://bugzilla.mozilla.org/show_bug.cgi?id=1348873)

<!-- Pull requests that do not address these steps are welcome, but they will require additional verification as part of the review process. -->

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/servo/servo/16047)
<!-- Reviewable:end -->

+++ components/style/gecko_bindings/structs_debug.rs
@@ -5026,7 +5026,7 @@ pub mod root {
             pub mStates: root::mozilla::EventStates_InternalType,
         }
         pub type EventStates_InternalType = u64;
-        pub type EventStates_ServoType = u16;
+        pub type EventStates_ServoType = u64;
         #[test]
         fn bindgen_test_layout_EventStates() {
             assert_eq!(::std::mem::size_of::<EventStates>() , 8usize , concat
